refactor(proactive): extract contracts and generation guards
